refractor positions table

+++ src/lib/components/table/TableActions.svelte
@@ -2,8 +2,8 @@
 	type Props = {
 		item: unknown;
 		canDelete?: boolean;
-		onEdit: (item: unknown) => void;
-		onDelete: (item: unknown) => void;
+		onEdit?: (item: unknown) => void;
+		onDelete?: (item: unknown) => void;
 		onView?: (item: unknown) => void;
 	};
 
@@ -17,9 +17,9 @@
 </script>
 
 <div class="flex justify-end gap-3">
-
 	{#if onView}
 		<button
+			type="button"
 			class="text-sm font-medium text-blue-700 hover:text-blue-900"
 			onclick={() => onView(item)}
 		>
@@ -27,21 +27,23 @@
 		</button>
 	{/if}
 
-	<button
-		class="text-sm font-medium text-black hover:text-zinc-700"
-		onclick={() => onEdit(item)}
-	>
-		Edytuj
-	</button>
-
-	{#if canDelete}
+	{#if onEdit}
 		<button
+			type="button"
+			class="text-sm font-medium text-black hover:text-zinc-700"
+			onclick={() => onEdit(item)}
+		>
+			Edytuj
+		</button>
+	{/if}
+
+	{#if onDelete && canDelete}
+		<button
+			type="button"
 			class="text-sm font-medium text-red-700 hover:text-red-900"
 			onclick={() => onDelete(item)}
 		>
 			Usuń
 		</button>
-	{:else}
-		<span class="text-sm text-zinc-500">Brak ID</span>
 	{/if}
 </div>
